+++ NeoForge/src/main/java/net/potionstudios/biomeswevegone/neoforge/datagen/generators/RecipeGenerator.java
@@ -10,6 +10,7 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.ItemLike;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.SoundType;
 import net.potionstudios.biomeswevegone.BiomesWeveGone;
 import net.potionstudios.biomeswevegone.tags.BWGItemTags;
@@ -198,6 +199,18 @@ public class RecipeGenerator extends RecipeProvider {
         dyeTagRecipe(recipeOutput, Items.PINK_DYE, BWGItemTags.MAKES_PINK_DYE);
 
         twoByTwoPackertoFourWithStoneCutting(recipeOutput, RecipeCategory.BUILDING_BLOCKS, BWGBlocks.RED_ROCK_BRICKS_SET.getBase(), BWGBlocks.RED_ROCK_SET.getBase());
+        ShapelessRecipeBuilder.shapeless(RecipeCategory.BUILDING_BLOCKS, BWGBlocks.MOSSY_RED_ROCK_BRICKS_SET.getBase())
+                .requires(BWGBlocks.RED_ROCK_BRICKS_SET.getBase())
+                .requires(Items.VINE)
+                .group("mossy_red_rock_bricks")
+                .unlockedBy(getHasName(Items.VINE), has(Items.VINE))
+                .save(recipeOutput, getConversionRecipeName(BWGBlocks.MOSSY_RED_ROCK_BRICKS_SET.getBase(), Items.VINE));
+        ShapelessRecipeBuilder.shapeless(RecipeCategory.BUILDING_BLOCKS, BWGBlocks.MOSSY_RED_ROCK_BRICKS_SET.getBase())
+                .requires(BWGBlocks.RED_ROCK_BRICKS_SET.getBase())
+                .requires(Items.MOSS_BLOCK)
+                .group("mossy_red_rock_bricks")
+                .unlockedBy(getHasName(Items.MOSS_BLOCK), has(Items.MOSS_BLOCK))
+                .save(recipeOutput, getConversionRecipeName(BWGBlocks.MOSSY_RED_ROCK_BRICKS_SET.getBase(), Items.MOSS_BLOCK));
         chiseled(recipeOutput, RecipeCategory.BUILDING_BLOCKS, BWGBlocks.CHISELED_RED_ROCK_BRICKS_SET.getBase(), BWGBlocks.RED_ROCK_BRICKS_SET.getBase());
         ShapelessRecipeBuilder.shapeless(RecipeCategory.BUILDING_BLOCKS, BWGBlocks.ROCKY_STONE_SET.getBase(), 2)
                 .requires(Items.COBBLESTONE)
@@ -208,9 +221,16 @@ public class RecipeGenerator extends RecipeProvider {
         ShapelessRecipeBuilder.shapeless(RecipeCategory.BUILDING_BLOCKS, BWGBlocks.MOSSY_STONE_SET.getBase())
                 .requires(Items.STONE)
                 .requires(Items.VINE)
-                .unlockedBy(getHasName(Items.STONE), has(Items.STONE))
+                .group("mossy_stone")
                 .unlockedBy(getHasName(Items.VINE), has(Items.VINE))
-                .save(recipeOutput);
+                .save(recipeOutput, getConversionRecipeName(BWGBlocks.MOSSY_STONE_SET.getBase(), Items.VINE));
+        ShapelessRecipeBuilder.shapeless(RecipeCategory.BUILDING_BLOCKS, BWGBlocks.MOSSY_STONE_SET.getBase())
+                .requires(Items.STONE)
+                .requires(Items.MOSS_BLOCK)
+                .group("mossy_stone")
+                .unlockedBy(getHasName(Items.MOSS_BLOCK), has(Items.MOSS_BLOCK))
+                .save(recipeOutput, getConversionRecipeName(BWGBlocks.MOSSY_STONE_SET.getBase(), Items.MOSS_BLOCK));
+
 
         smeltingResultFromBase(recipeOutput, BWGBlocks.DACITE_SET.getBase(), BWGBlocks.DACITE_COBBLESTONE_SET.getBase());
         twoByTwoPackertoFourWithStoneCutting(recipeOutput, RecipeCategory.BUILDING_BLOCKS, BWGBlocks.DACITE_BRICKS_SET.getBase(), BWGBlocks.DACITE_SET.getBase());