+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.1.5
+- Fix and Add Missing Mossy Red Rock and Mossy Stone Recipes
 
 # 2.1.4 (Forge Only)
 - Require Forge 52.0.20 as a minimum due to the addition of new Forge tags and implementation of Common Tags